requirements.txt
pandas
seaborn
matplotlib
azure-storage-blob
azure-functions